FIXED DEPTH ERROR
I originally wrote this algorithm in C++, but was had the issue that the test case for depth 24 was being found at depth 26.  I tried for at least a week to fix this bug, but I could not for the life of me find the reason that it was occurring, and thus, could not fix it. I finally figured out that I was marking the nodes as visited when they were inserted into the queue, instead of when they were removed, causing more optimal paths to be ignored. I thought that if I marked them as visited when I inserted them, I would avoid inserting duplicate states into the queue, but that appears to be the entire point of the algorithm. I have now fixed the bug in both my python implementation and my C++ implementation (which runs MUCH, MUCH faster by the way) and am very happy.